chore(registry): move to `api/internal` namespace
