+++ api/internal/registry/registry_test.go
